\name{ohiovowels}
\docType{data}
\alias{ohiovowels}
\title{ ohiovowels: Sample data for vowels.R }
\description{
  A sample dataset for examples for vowels.R. 
}
\usage{
data(ohiovowels)
}
\format{A data frame containing 447 rows and 9 columns}
\details{
This dataset includes 477 vowels from 2 speakers, a female and male, from Ohio.  This is the "CentralOhioNORM.txt" file available on the NORM website.
}
\source{\url{http://lingtools.uoregon.edu/norm/downloads/CentralOhioNORM.txt}}
\author{ Thomas, Erik <erthomas@ncsu.edu> }
\keyword{datasets}